setbrightness action fix

+++ components/ehmtx/__init__.py
@@ -178,14 +178,13 @@ SET_BRIGHTNESS_ACTION_SCHEMA = cv.Schema(
 
 SetBrightnessAction = ehmtx_ns.class_("SetBrightnessAction", automation.Action)
 
-
 @automation.register_action(
     "ehmtx.set.brightness", SetBrightnessAction, SET_BRIGHTNESS_ACTION_SCHEMA
 )
 async def ehmtx_set_brightness_action_to_code(config, action_id, template_arg, args):
     paren = await cg.get_variable(config[CONF_ID])
     var = cg.new_Pvariable(action_id, template_arg, paren)
-    template_ = await cg.templatable(config[CONF_BRIGHTNESS], args, cg.int32)
+    template_ = await cg.templatable(config[CONF_BRIGHTNESS], args, cg.int_)
     cg.add(var.set_brightness(template_))
 
     return var
